src/main/java/com/biperf/core/domain/enums/ClaimStatusType.java

package com.biperf.core.domain.enums;

import java.util.List;

/**
 * ClaimStatusType.
 * <p>
 * <b>Change History:</b><br>
 * <table border="1">
 * <tr>
 * <th>Author</th>
 * <th>Date</th>
 * <th>Version</th>
 * <th>Comments</th>
 * </tr>
 * <tr>
 * <td>robinsra</td>
 * <td>Nov 22, 2005</td>
 * <td>1.0</td>
 * <td>created</td>
 * </tr>
 * </table>
 * 
 *
 */
public class ClaimStatusType extends PickListItem
{

  /**
   * Asset name used in Content Manager
   */
  public static final String PICKLIST_ASSET = "picklist.claim.status.type";

  public static final String OPEN = "open";
  public static final String CLOSED = "closed";

  /**
   * This constructor is protected - only the PickListFactory class creates these instances.
   */
  protected ClaimStatusType()
  {
    super();
  }

  /**
   * Get the pick list from content manager.
   * 
   * @return List
   */
  public static List getList()
  {
    return getPickListFactory().getPickList( ClaimStatusType.class );
  }

  /**
   * Returns null if the code is null or invalid for this list.
   * 
   * @param code
   * @return PickListItem
   */
  public static ClaimStatusType lookup( String code )
  {
    return (ClaimStatusType)getPickListFactory().getPickListItem( ClaimStatusType.class, code );
  }

  /**
   * Returns null if the defaultItem is not defined - or invalid
   * 
   * @return default PickListItem
   */
  // public static ClaimStatusType getDefaultItem()
  // {
  // return (ClaimStatusType)getPickListFactory().getDefaultPickListItem( ClaimStatusType.class
  // );
  // }
  /**
   * Overridden from
   * 
   * @see com.biperf.core.domain.enums.PickListItem#getPickListAssetCode()
   * @return PICKLIST_ASSET
   */
  public String getPickListAssetCode()
  {
    return PICKLIST_ASSET;
  }

}
